Calculate Feed Conversion Ratio (FCR)

Calculates Feed Conversion Ratio (FCR) based on Average Daily Gain (ADG) and Average Daily Feed Intake (ADFI) data. This function merges ADG and ADFI data, calculates corrected FCR values, and generates statistical summaries.

fcr_get(adg_res, adfi_res, quiet = FALSE)

Arguments

adg_res

A list object returned by the adg_get() function, containing ADG data and related information.

adfi_res

A list object returned by the adfi_get() function, containing ADFI data and related information.

quiet

Logical; if TRUE, suppresses all messages and warnings, default is FALSE.

Value

A list containing:

fcr_res: A data.table with FCR calculation results

adg_info: The input ADG information table

adg_data: The original ADG data

adfi_info: The input ADFI information table

adfi_data: The original ADFI data

fcr_summary: Statistical summary of FCR calculation results

Details

The fcr_get function performs the following steps:

Validates the format of input ADG and ADFI data

Merges ADG and ADFI data, matching by animal ID and location

Calculates corrected daily gain and feed conversion ratio

Generates statistical summaries of FCR, including sample size, minimum, maximum, median, mean, standard deviation, and coefficient of variation

See also

Examples

result_nedap <- preprocess_data(data = mintyr::nedap, station_type = "nedap", quiet = TRUE)
adg_result <- adg_get(result_nedap, my_break = c(30,120), quiet = TRUE)
adfi_result <- adfi_get(result_nedap, adg_result, quiet = TRUE)
fcr_result <- fcr_get(adg_res = adg_result, adfi_res = adfi_result, quiet = TRUE)
head(fcr_result$fcr_res)
#> Key: <responder, location>
#>    responder location  stage start_date_cut min_weight_cut end_date_cut
#>       <char>   <char> <char>         <Date>          <num>       <Date>
#> 1:     13913      101 30-120     2024-03-02       30500.00   2024-05-16
#> 2:     13918      101 30-120     2024-03-07       29846.15   2024-05-24
#> 3:     13935      102 30-120     2024-02-28       31388.89   2024-05-09
#> 4:     13954      101 30-120     2024-02-25       35250.00   2024-05-23
#> 5:     13996      101 30-120     2024-02-27       32136.36   2024-05-06
#> 6:     14260      102 30-120     2024-02-23       31437.50   2024-05-07
#>    max_weight_cut  lm_slope r_squared stage_days      adg origin_dfi
#>             <num>     <num>     <num>      <num>    <num>      <num>
#> 1:       121171.9 1193.6622 0.9938655         75 1200.000   2535.959
#> 2:       119208.3 1147.9651 0.9968953         78 1153.846   2319.814
#> 3:       119937.5 1261.9862 0.9973985         71 1267.606   2703.371
#> 4:       112295.5  998.5019 0.9375183         88 1022.727   2317.260
#> 5:       119850.0 1308.1903 0.9950834         69 1304.348   2909.309
#> 6:       121000.0 1224.6884 0.9954987         74 1216.216   2846.546
#>    corrected_dfi corrected_adg  cor_fcr
#>            <num>         <num>    <num>
#> 1:      2560.189      1196.831 2.139140
#> 2:      2307.572      1150.906 2.005005
#> 3:      2704.603      1264.796 2.138371
#> 4:      2307.228      1010.615 2.282995
#> 5:      2921.705      1306.269 2.236679
#> 6:      2822.715      1220.452 2.312844
head(fcr_result$fcr_summary)
#>            traits      N     min     max  median    mean     sd     cv
#>            <char> <char>  <char>  <char>  <char>  <char> <char> <char>
#> 1: min_weight_cut     28   27.80   35.25   31.13   31.35   1.40  4.46%
#> 2: max_weight_cut     28  104.16  121.38  119.49  118.48   3.48  2.94%
#> 3:       lm_slope     28  934.27 1358.17 1219.19 1199.13 107.21  8.94%
#> 4:     stage_days     28   66.00   96.00   74.00   75.36   7.53 10.00%
#> 5:            adg     28  937.50 1363.64 1216.22 1204.68 108.19  8.98%
#> 6:  corrected_dfi     28 2039.54 3333.84 2817.13 2745.49 333.44 12.14%