Import Nedap Data CSV Files

Imports and combines data from multiple Nedap data csv files using either the data.table or vroom package. This function provides a flexible way to read and merge multiple data files with consistent formatting.

import_nedap(filepaths, package = "data.table", ...)

Arguments

filepaths

A character vector of file paths to the Nedap data files to be imported.

package

Character string specifying which package to use for importing; either "data.table" or "vroom". Default is "data.table".

...

Additional arguments passed to the underlying import function (data.table::fread or vroom::vroom).

Value

A data.table or tibble (depending on the package used) containing the combined data from all imported files.

Details

The import_nedap function performs the following steps:

Validates that all specified file paths exist

Checks that the requested package is installed and supported

Imports all files using the specified method

Combines all imported data into a single data structure

When using "data.table" as the package, files are read individually with fread() and combined with rbindlist(). When using "vroom", all files are read at once using vroom's multi-file capability.

See also

Examples

# Example: CSV file import demonstrations

# Setup test files
csv_files <- mintyr::mintyr_example(
  mintyr::mintyr_examples("csv_test")     # Get example CSV files
)

import_nedap(filepaths = csv_files)
#> ℹ Using package data.table to import 2 files.
#> ✔ Successfully imported data from 2 files
#>     col1   col2   col3
#>    <int> <char> <lgcl>
#> 1:     4      d  FALSE
#> 2:     5      f   TRUE
#> 3:     6      e   TRUE
#> 4:    15      o  FALSE
#> 5:    16      p   TRUE
#> 6:    17      q  FALSE